Refactor: CLI to switch seamlessly jump betewen prompt and and flags
The CLI will now ask via promopt for all missing flags.
This was already the case, however, the CLI would ask for email
and password although the account might be set to public.
This version has a call back that assess which prompts are necessary
and which are not given the account to scrape from is public or nor.